requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-recaptcha==4.0.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-template-partials==23.4
django-use-email-as-username==1.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
idna==3.6
mypy==1.9.0
mypy-extensions==1.0.0
phonenumbers==8.13.33
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
urllib3==2.2.1
